# core/services/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/aws/iam/groups": {
            "get": {
                "description": "Get a list of AWS IAM groups",
                "produces": [
                    "text/plain"
                ],
                "summary": "AWS List Groups",
                "operationId": "aws-list-groups",
                "responses": {
                    "200": {
                        "description": "Group names separated by newline",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new AWS IAM group",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create AWS Group",
                "operationId": "create-aws-group",
                "parameters": [
                    {
                        "description": "Group details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/iam.CreateGroupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gitea_pcp-inariam_inariam_core_services_api_responses_aws_iam.Group"
                        }
                    }
                }
            }
        },
        "/aws/iam/groups/{groupName}": {
            "get": {
                "description": "Get AWS IAM group details by group name",
                "produces": [
                    "application/json"
                ],
                "summary": "Get AWS Group by Name",
                "operationId": "get-aws-group-by-name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group Name",
                        "name": "groupName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gitea_pcp-inariam_inariam_core_services_api_responses_aws_iam.Group"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing AWS IAM group's name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update AWS Group",
                "operationId": "update-aws-group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group Name",
                        "name": "groupName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New group details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gitea_pcp-inariam_inariam_core_services_api_requests_aws_iam.UpdateGroupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gitea_pcp-inariam_inariam_core_services_api_responses_aws_iam.Group"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an AWS IAM group by group name",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete AWS Group",
                "operationId": "delete-aws-group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group Name",
                        "name": "groupName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "AWS Group deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/policies/{policyName}": {
            "get": {
                "description": "Get IAM policy details by policy name",
                "produces": [
                    "application/json"
                ],
                "summary": "Get Policy by Name",
                "operationId": "get-policy-by-name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Policy Name",
                        "name": "policyName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gitea_pcp-inariam_inariam_core_services_api_responses_aws_iam.PolicyDetailResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an IAM policy by policy name",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete Policy",
                "operationId": "delete-policy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Policy Name",
                        "name": "policyName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Policy deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "gitea_pcp-inariam_inariam_core_services_api_requests_aws_iam.UpdateGroupRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "new_name": {
                    "type": "string"
                }
            }
        },
...
        "gitea_pcp-inariam_inariam_core_services_api_responses_aws_iam.Group": {
            "type": "object",
            "properties": {
                "arn": {
                    "type": "string"
                },
                "created_date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        },
        "gitea_pcp-inariam_inariam_core_services_api_responses_aws_iam.PolicyDetailResponse": {
            "type": "object",
            "properties": {
                "policy_id": {
                    "type": "string"
                },
                "policy_name": {
                    "type": "string"
                }
            }
        },
...
        "iam.CreateGroupRequest": {
            "type": "object",
            "required": [
                "description",
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    }
}